+++ resources/views/components/emails/footer-plain.blade.php
@@ -5,6 +5,10 @@
 @lang('ig-common::layouts.email.generator', ['generator' => InternetGuru\LaravelCommon\Support\Helpers::getAppInfo()])
 
 @lang('ig-common::layouts.email.requested-from', ['ip' => $ip, 'timezone' => $timezone ?? 'n/a'])
+@if ($noreplyMessage)
+
+{{ $noreplyMessage }}@endif
 
 
 @lang('ig-common::layouts.provider', ['link' => __('ig-common::layouts.provider.www'), 'year' => date('Y')])
+