+++ src/Mail/MailMessage.php
@@ -7,7 +7,9 @@ use Illuminate\Support\Str;
 
 class MailMessage extends BaseMailMessage
 {
-    protected array $extraMailData = [];
+    protected array $extraMailData = [
+        'noreplyMessage' => '',
+    ];
 
     protected string $refNumber;
 
@@ -31,15 +33,44 @@ class MailMessage extends BaseMailMessage
         return $this;
     }
 
+    public function replyTo($address, $name = null)
+    {
+        $this->updatenoreplyMessage($address);
+
+        return parent::replyTo($address, $name);
+    }
+
+    public function from($address, $name = null)
+    {
+        $this->updatenoreplyMessage($address);
+
+        return parent::from($address, $name);
+    }
+
     public function view($view, array $data = [])
     {
-        return parent::view($view, $data + $this->extraMailData);
+        $address = $this->from[0][0] ?? config('mail.from.address', '');
+        $this->updatenoreplyMessage($address);
+
+        return parent::view($view, array_merge($data, $this->extraMailData));
     }
 
     public function setExtraMailData(array $data): static
     {
-        $this->extraMailData = $data;
+        $this->extraMailData = array_merge($this->extraMailData, $data);
 
         return $this;
+    }
+
+    private function updatenoreplyMessage(?string $address): void
+    {
+        if (! $address) {
+            return;
+        }
+        if (Str::contains($address, ['no-reply', 'noreply'], ignoreCase: true)) {
+            $this->setExtraMailData(['noreplyMessage' => __('ig-common::layouts.email.no-reply-note')]);
+        } else {
+            $this->setExtraMailData(['noreplyMessage' => '']);
+        }
     }
 }