+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+### Added
+
+- Add automatic replyto message to email footers when from / replyto address is no-reply.
+
 ## [4.6.4] - 2026-01-25
 
 ### Fixed
@@ -1140,6 +1146,7 @@ _Stable release based on [0.1.0-rc.1]._
 
 ## [0.0.0] - 2024-09-12
 
+[Unreleased]: https://https://github.com/internetguru/laravel-common/compare/staging...dev
 [4.6.4]: https://https://github.com/internetguru/laravel-common/compare/v4.6.3...v4.6.4
 [4.6.3]: https://https://github.com/internetguru/laravel-common/compare/v4.6.2...v4.6.3
 [4.6.2]: https://https://github.com/internetguru/laravel-common/compare/v4.6.1...v4.6.2